Add new NUCLEO-F030R8 variant
This is a first pass at getting the NUCLEO-F030R8 board working.
I've tested the digital pins, the PWM code using tone and the
Serial USART.  To get the code to compile I hacked up analog.c
to avoid using the DAC functions.  This is pretty ugly. I will
probably rework it.