Fix SPI transfer16
When the SPI device is configured for MSB first, the 16 bit data being
sent and received on the SPI bus has the bytes swapped if the
SPI.transfer16() method is used. Swapping the bytes before and after the
call to spi_transfer() fixes the problem.

Tested using an AS5147P magnetic encoder spi device. Before the fix, saw
frame errors due to the bytes being swapped, when the same unmodified
code worked perfectly on an AVR based SparkFun RedBoard.

Signed-off-by: Theodore A. Roth <troth@openavr.org>