Make Serialx instance generic
Previously Serialx instance was defined by the variant.
It was not enough generic and flexible.

With this patch, all available Serialx instance are handled dynamically.
Where 'x' is the U(S)ART number.
Currently, STM32 MCU could have 9 U(S)ART.

Serial is the generic name used in Arduino sketches.
Serial is now mapped to the Serialx instance linked
to the com port in the variant.h (mainly one linked to ST-Link)
Moreover this will simplify USB integration to map the SerialUSB.

Signed-off-by: Frederic.Pillon <frederic.pillon@st.com>