Some linker script has the stack pointer __NOT__ 8-byte aligned
Ex: 0x20001FFF is the last valid byte address of RAM for NUCLEO_L053R8,
but the error is that it is not aligned and there is no reason to put
this value as stack pointer.
When Cortex-M push something in the stack, it first decrement the stack
pointer and then writes at the new address.
So using initial value 0x20002000 is ok.
The first word will be at address _estack-4, in RAM

Signed-off-by: Frederic.Pillon <frederic.pillon@st.com>