Allow to redefine Exception handler
By default, all Interruption and Exception Handlers are weaked and
aliased to Default_Handler (Infinite Loop, See CMSIS device startup file)
So, those defined handlers in the core doing a loop bring nothing
except to not allow further redefinition.
But this is what we want, be able to redefine those handler.
Ex: FreeRTOS library requires to be able to redefine them.

Signed-off-by: Frederic.Pillon <frederic.pillon@st.com>