Removed printf call in _Error_Handler() to reduce code size
By using the printf function of the library C this inflates
the size of the code. An alternative, will be to implement
a tiny and limited functionality implementation of printf.

Example, build modified Blink.ino with _Error_Handler()
calls for Nucloe F103RB:
With printf: Sketch uses 16368 bytes
Without printf: Sketch uses 10220 bytes
Difference is 6k of flash used which is not negligible.
Nano static library is linked to use this printf call.

Signed-off-by: Frederic.Pillon <frederic.pillon@st.com>