Rework eeprom emulation for buffered access
The current eeprom emulation didn't had any buffered access to the
data on the flash. Every access to a byte would trigger a whole
page read / write. For writes it would mean, that the whole page
would be erased. For backward compability we will keep
the current functions as they are, but will add functions
to fill/flush the buffer and read from it.

The buffer size is static and is one page in size.

This commit also cleans up old unused code comments and
documentation stubs. I also cleaned up the old references
to the original author / version / date, since we already
refactored this code quite well. The copyright notice of
course is still intact.

Fixes: #296