[EEPROM emulation] Rework for buffered access
The current EEPROM emulation didn't had any buffered access to the
data on the flash. Every access to a byte would trigger a whole
page read / write. For writes it would mean, that the whole page
would be erased. For backward compability we will keep
the current functions as they are, but will add functions
to fill/flush the buffer and read from it.

The buffer size is static and is one page in size.

Fixes: #296

Signed-off-by: Nils Hasenbanck <nils.hasenbanck@trendig.com>