Fix SPI::setClockDivider implementation
As said in Arduino API:
https://www.arduino.cc/en/Reference/SPISetClockDivider

"On the Due, the system clock can be divided by values
from 1 to 255. The default value is 21, which sets the clock
to 4 MHz like other Arduino boards."

So in fact, it should be possible to divide the clock by any value.
Then HAL will use the best prescaler to fit requested Freq.

Moreover, it is not possible to define SPI_CLOCK_DIVx to match
the 16MHz value for avr as SPI frequency depends of system clock
configuration.

Fix #299

Signed-off-by: Frederic.Pillon <frederic.pillon@st.com>