Allow to change default PWM frequency
Default is PWM_FREQUENCY (1000Hz).
Call `analogWriteFrequency(freq)` before `analogWrite(pin, val)`
to use the new frequency.

Note that no compute/check is performed to ensure the requested freq.
This is dependent on how the computation is currently done in analogWrite.
This limitation will be remove when Timer management will be done.

Signed-off-by: Frederic.Pillon <frederic.pillon@st.com>