Fix SPI pin matching check
This looks up the SPI module corresponding to each of pins passed and
then uses `pinmap_merge_peripheral()` to merge this in a single pin. If
not all pins map to the same SPI module, this should abort.

The `pinmap_merge_peripheral` function returns NP to indicate a mismatch
between the passed peripherals. However, when one of the arguments is
already NP, this is not propagated into the function result, instead the
other argument is simplly returned.

In this case, the merging happens in two levels: First MISO and MOSI are
merged, then SCK and SS are merged and finally the merged results are
again merged. That means that if MISO and MISO mismatch, but SCK and SS
match (or the other way around), the NP returned due to the mismatch is
ignored and the final result is not NP, so no error is raised.

This changes the code to also check the intermediate merge results for
NP, to properly detect mismatches.