[I2C] Fix error reporting on I2C timeout
When this timeout is triggered, I2C_OK was returned. This happens because in the
loop above, it checks for `(delta < I2C_TIMEOUT_TICK)`, so it breaks out of the
loop as soon as `delta == I2C_TIMEOUT_TICK`. Then, the if below checks for
`(delta > I2C_TIMEOUT_TICK)`, which is not yet false. Since the hardware reports
no errors either, I2C_OK was returned.

By using `>=` in the if rather than `>`, this is fixed and I2C_TIMEOUT is now
returned instead.