HardwareTimer: Enable preload for ARR/overflow register
This changes the default for preload (i.e. delay updating the overflow
value until an update event such as timer overflow) to enabled, since
this is will probably make the timer output behave as expected as much
as possible (e.g. you cannot make it "skip" an overflow and run it all
the way up to 0xffff by lowering the overflow value).