Support a .noinit section for variables
This is inspired by the linker sripts on the AVR architecture, which
support a .noinit section (or any section starting with .noinit,
actually) for variables that should be allocated an address in RAM, but
not be initialized to any particular value (not even zero) on startup.
These can then be used to remember values across resets.

From the sketch perspective, this works exactly the same as on AVR: Just
annote a global variable with `__attribute__((__section__(".noinit")))`
and it will have an unpredictable value on power-up and retain its value
during resets.

To implement this without having to change all board-specific linker
scripts, the linker commandline is changed to pass the board-specific
linker script to the `--default-script` linker script, and change the
main linker script (passed to `--script`, previously `-T`) to a generic
"override" linker script. This new generic linker script contains an
`INSERT BEFORE` command, which causes the linker to load it *in addition
to* the default linker script, while adding an extra `.noinit` output
section in the right place.

Because these new variables take up RAM but have their own section in
the .elf file, they should be accounted for in the size summary after
compilation. This is done by adapting the `recipe.size.regex.data` entry
to include this new section.