Do not use the RTC to track a bootloader request
Instead, use a .noinit variable, which survives across reboots.

Note: This file uses the bool type, so needs to include stdbool.h.
It seems that for the F4, this was indirectly included elsewhere,
but for e.g. F0, this did not seem to be the case, causing the build
to fail.