Use handcoded delay in USBD_reenumerate
This now runs in early startup, so the normal delay will not work
(systick is not running yet).

Note: Seems that on F0, inline assembler is compiled using the classic
rather than unified syntax (which *is* used on F4 it seems). This
explicitly selects the unified syntax in the inline assembly block, to
fix building for F0.

TODO: There is a variant which also overrides the enumeration.