Simplify system flash address detection
On most CPUs, detecting the system flash address is not needed, since
they can remap the system flash onto adress 0x0 and the bootloader reset
vector can be read from there. This approach also more closely mimics a
bootloader start using the BOOTx pins.

For CPUs that cannot remap flash, this hardcodes the address of system
flash and jumps there instead. For H7 and F7, this is also how it works
when using the BOOT pins. For F1, remapping can be done using the BOOT
pins, but not from software, resulting in a difference with a
BOOT-activated bootloader.

This puts the code that figures out where the bootloader lives back in a
separate function. This makes the code easier to read, and makes it
easier to jump to custom bootloaders (e.g. in normal flash instead of in
system flash).

TODO: Check value for H7, AN2606 and reference manual disagree