Refactor USB pullup handling
Previously, variants could define USB_DISC_PIN when they had an USB
attach pullup that was disabled by default and needed a pin to
be written LOW to enable it. Other hardware configurations could only
overwrite the USBD_reenumerate function, like the M200 board did.

This commit makes the pullup configuration more flexible. By defining
the appropriate macros, enabled-by-default and disabled-by-default
pullups are both supported. The output level of the pin can also be
configured.

The code that controls an external pullup is now merged with the code
that applies the D+ output trick for fixed pullups, since the behaviour
is almost identical (except for reverting to INPUT mode for the D+ pin
instead of inverting the output value).

Finally, for CPUs that have internal USB pullup (as indicated by the
presence of a SDIS or DPPU config bit), the write-to-DP-trick is now not
performed, but the internal pullup is toggled instead.

This fixes #885, also see that issue for discussion leading up to this
change.