WIP: Allow using "Burn bootloader" in the Arduino IDE
This allows uploading the HID bootloader and the stm32duino bootloader
(aka maple bootloader 2.0) to the bluepill and blackpill boards (other
boards could also be added, but a bootloader version must be explicitly
selected for each board).

The bootloader is chosen based on the "Upload method" selected and
uploaded with the selected programmer using openocd.

This uses the non-fastboot bootloader, since the current bootloader
autoreset for this bootloader just does a normal reset. If that would be
modified to set the right RTC register with a magic value to stop the
fastboot in the bootloader (pretty much exactly like the hid
bootloader), we could instead use the fastboot version.