Added STM32F072CB varant for the Elektor LoRa Node
As the STM32F072C8 ( 64kB ) is not proviiding much space for user applications after the LoRaWAN Stack is inside, the first production run will include a STM32F072CB ( 128kB) with the same pinout but a bit more falsh. The Variant has been renamed to reflect this change, also a new Board has added to the menu so you can choose what you have populated.