HardwareTimer: pause() need to call HAL API to restore HAL state
Due to HAL Cube update (ex: for STM32L4 sha1 f41f10e)
it is now required to stop timer with HAL API
otherwise HAL state is not restored to HAL_TIM_STATE_READY,
and it is not possible to resume() a HardwareTimer after a pause().
Nevertheless it is not sufficient to guarantee that timer is stopped
specially if some channels are still running.
So it is also necessary to keep also the call to
LL_TIM_DisableCounter()

Signed-off-by: Alexandre Bourdiol <alexandre.bourdiol@st.com>