feat: easy_setup prompts board codenames
Using a fuzzy find after parsing boards.txt.

This has required refactoring the board parser code in a dedicated modules,
with minor impacts in the files that depended on it.

=============

Doc changes:

expand on the syntax in easy_setup:
Added a comment block to explain the use of the DEPENDS clause.
Added "usual" values for board settings.
Add a link to the wiki in easy_setup.

Also update link in the README_CMAKE
- update link to point to the new location for the examples;
- unify the displayed text for the links;
- add a link to the wiki.