feat: Add SubGhz library to access SubGhz module
This allows accessing some of the signals internally connected to the
SubGhz radio (that would have been GPIO signals if the radio was
external). This also allocates and exposes the SPI object connected to
the SubGhz radio block and handles attaching handlers to the radio
interrupt.

Note that the DIO signals are *not* exposed, since there is no way to
read them directly (and indirectly reading them through the IRQ pending
flag does not work in all cases).