feat(SPI): Make SPISettings constructors constexpr
This turns SPISettings into a "literal type" and allows variables using
it to be constexpr to guarantee the constructor is executed at
compiletime. This requires (with C++11) that all variables are set using
initializers instead of assignments and that the regular if cascade is
replaced by a ternary if. It also requires explicit initializers for all
values (omitted variables were previously initialized to zero anyway).