Fix links for "WeAct MiniSTM32H7*" boards
The previous links lead through a GitHub username change redirect to a repository that appears to be used to demonstrate a security vulnerability.

I replaced them with links to the repo under the WeActStudio GitHub account which actually contains information about the hardware. The other links for WeAct hardware in the readme are under that GitHub and the product listings under the "WeAct Studio Official Store" on AliExpress.