fix: silence the warnings from ExternalProject_Add
CMake 3.24 introduced a change in behavior in ExternalProject_Add()
regarding timestamps when extracting an archive.
As updating the code would break compatibility with older CMake versions,
this commit instead tells newer CMake versions to behave like older ones,
using a policy setting.

Note from the CMake docs:
The OLD behavior of a policy is deprecated by definition and may be removed in a future version of CMake.

-> so the code will have to be updated eventually, and compatibility be broken...